Google function code/requirements.txt
scikit-learn==1.2.2
numpy==1.21.2
pandas==1.3.3
google-cloud-storage==1.44.0